refactor(ai): move tool bridge into @epicenter/workspace/ai and redesign API
@epicenter/ai was a 213-LOC package that didn't earn its own package.json—
it only bridged workspace actions to TanStack AI types, and its sole
dependency was @epicenter/workspace. Moving it into a workspace subpath
export eliminates the package overhead while keeping the AI bridge
as an opt-in import that doesn't pollute the main workspace entry point.

The API is redesigned around a single-object return:

  const aiTools = actionsToAiTools(workspace.actions);
  aiTools.tools        // AnyClientTool[] for ChatClient
  aiTools.definitions  // wire-safe ToolDefinition[] for the server

Each definition includes the action's title when declared, so consumers
can show human-readable labels without re-walking the action tree.

This replaces the old two-function pipeline (actionsToClientTools then
toToolDefinitions). Title is now forwarded to definitions, fixing a dead
field on the old ToolDefinitionPayload that was declared but never
populated. The description fallback now uses underscores consistently
(it previously used dots for descriptions but underscores for names).
