+++ apps/opensidian/package.json
@@ -18,7 +18,6 @@
 	},
 	"dependencies": {
 		"@codemirror/autocomplete": "^6.20.1",
-		"@epicenter/ai": "workspace:*",
 		"@epicenter/constants": "workspace:*",
 		"@epicenter/skills": "workspace:*",
 		"@epicenter/svelte": "workspace:*",
