+++ apps/opensidian/src/lib/client.ts
@@ -1,4 +1,4 @@
-import { actionsToClientTools, toToolDefinitions } from '@epicenter/ai';
+import { actionsToAiTools } from '@epicenter/workspace/ai';
 import { APP_URLS } from '@epicenter/constants/vite';
 import { createSqliteIndex, createYjsFileSystem } from '@epicenter/filesystem';
 import { createSkillsWorkspace } from '@epicenter/skills';
@@ -213,31 +213,11 @@ export const auth = createAuth({
 	},
 });
 
-/**
- * Workspace actions converted to TanStack AI client tools.
- *
- * Each action becomes a client-side tool with `__toolSide: 'client'` and
- * its handler wired as the `execute` function. Queries auto-execute;
- * mutations trigger the approval UI.
- *
- * Passed to `createChat({ tools: workspaceTools })` for local execution.
- */
-export const workspaceTools = actionsToClientTools(workspace.actions);
+/** AI tool representations for the opensidian workspace. */
+export const workspaceAiTools = actionsToAiTools(workspace.actions);
 
-/**
- * Wire-safe tool definitions stripped for the HTTP request body.
- *
- * Removes `execute` functions (not JSON-serializable) and normalizes
- * input schemas for provider compatibility (Anthropic requires
- * `properties` + `required` on every schema).
- *
- * Sent to the server in `body.data.tools` so `chat()` knows what
- * tools exist without needing them hardcoded.
- */
-export const workspaceDefinitions = toToolDefinitions(workspaceTools);
-
-/** All workspace tool names as a type union. */
-export type WorkspaceTools = typeof workspaceTools;
+/** Tool array type for use in TanStack AI generics. */
+export type WorkspaceTools = typeof workspaceAiTools.tools;
 
 /** Yjs-backed virtual filesystem with path-based operations. */
 export const fs = createYjsFileSystem(
