--- packages/ai/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "@epicenter/ai",
-	"version": "0.1.0",
-	"private": true,
-	"type": "module",
-	"main": "./src/index.ts",
-	"types": "./src/index.ts",
-	"exports": {
-		".": "./src/index.ts"
-	},
-	"scripts": {
-		"typecheck": "tsc --noEmit"
-	},
-	"dependencies": {
-		"@epicenter/workspace": "workspace:*",
-		"@tanstack/ai": "catalog:"
-	},
-	"devDependencies": {
-		"typescript": "catalog:"
-	},
-	"license": "MIT"
-}
