--- packages/ai/src/index.ts
@@ -1,6 +1,0 @@
-export {
-	type ActionNames,
-	actionsToClientTools,
-	type ToolDefinitionPayload,
-	toToolDefinitions,
-} from './tool-bridge';
