+++ packages/workspace/src/ai/index.ts
@@ -1,0 +1,1 @@
+export { actionsToAiTools, type ToolDefinition } from './tool-bridge';
